src/service-b-python/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
grpcio==1.57.0
grpcio-tools==1.57.0
prometheus-client==0.17.1
python-json-logger==2.0.7
structlog==23.1.0
